Streamline SuperUser handling
The SuperUser is now no longer excluded from e.g. storing channel
listeners in the DB. This should now make the handling of this special
user more consistent with other parts of the code.
Additionally, the SuperUser was handled specially only by chance as the
user ID was checked to be != 0 to exclude the SuperUser. However, in
general the Mumble::SUPERUSER_ID could have any arbitrary value.