FIX(client): Fix crash on mono pos. audio warning from audio thread
Merge request #5247 introduced a warning that is triggered
when Mumble is configured to use positional audio and
only a single channel output is configured.

However, this warning is emitted from the audio thread and
therefore causes a crash. This is because Qt does not want to
add child QObjects from any other thread than main.

This commit makes sure the static version of ``logOrDefer`` is
called and also adds a check to ``Log::log`` to forward the call
to the main thread, if necessary.

Fixes #6507